规整Stream buffer read/write接口; 去除LLBC_MemCpy/LLBC_MemSet宏

+++ llbc/src/comm/Session.cpp
@@ -171,7 +171,7 @@ int LLBC_Session::Send(LLBC_MessageBlock *block)
         sessionSndBufUsed += _socket->GetIocpSendingDataSize();
     #endif
 
-    if (_sessionOpts.GetSessionSendBufSize() != LLBC_INFINITE &&
+    if (_sessionOpts.GetSessionSendBufSize() != static_cast<size_t>(LLBC_INFINITE) &&
         (sessionSndBufUsed + block->GetReadableSize()) >= _sessionOpts.GetSessionSendBufSize())
     {
         LLBC_Recycle(block);
